style: 💄 responsive redone
