+++ src/components/organisms/Footer.pcss
@@ -1,9 +1,5 @@
 footer {
-  @apply w-full p-4 text-xs text-center text-white bg-black/70;
-
-  /* .legal {
-    @apply ;
-  } */
+  @apply w-full md:w-2/3 mx-auto p-4 text-xs text-center text-white bg-black/70 rounded-t-lg;
 
   .layout {
     a {
