+++ src/routes/Single.pcss
@@ -1,32 +1,8 @@
-/* .categories-wrapper,  */.single-event--cover, .single-event--description {
-  @apply mb-5;
-}
-
 .single-event--wrapper {
-  @apply mx-auto w-5/6;
+  @apply mx-auto lg:w-5/6;
 
   .section-title {
     @apply ml-0;
   }
-}
-
-.single-event--header {
-  @apply flex items-baseline justify-between my-5;
-
-  .date {
-    @apply text-xl;
-  }
-}
-
-.categories-wrapper {
-  @apply flex gap-2;
-}
-
-.single-event--cover {
-  @apply mx-auto max-w-full rounded-md;
-}
-
-.single-event--description {
-  @apply whitespace-pre-wrap text-lg;
 }
