+++ src/routes/old.Portal.tsx
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import { useEffect, useState, useRef } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 import { AxiosResponse } from 'axios';
