+++ src/routes/old.index.tsx
@@ -1,3 +1,4 @@
+// @ts-nocheck
 import {
   createBrowserRouter, createRoutesFromElements, Route,
   RouterProvider,
